refactor(OUT-3460): separate logic from the editor component
- [x] create dedicated hook for handling the editor logic (extensions, conversion utils, sync, key handling)
